requirements.txt
streamlit
transformers
sentence-transformers
faiss-cpu
langchain
tiktoken
openai
chromadb  
nltk
spacy
python-dotenv
pandas
scikit-learn
notebook
ipykernel
PyPDF2
unstructured
pdfminer.six
accelerate 
bitsandbytes